GSEAdv: A package for analyzing gene sets.

The GSEAdv package provides methods enhancing the GSEABase package. Mainly it address thre points related to GeneSetCollections, describe, analyze, and simulate them. Each one has a specific vignette.

GSEAdv methods

There are several new methods for GeneSetCollections

summary: Some statistics of the GeneSetCollection.

genesPerPathway: The number of genes each gene set has.

pathwaysPerGene: The number of gene sets each gene has.

pathway: Information about a gene set.

gene: Information about a gene.

condPerGenes: The conditional probability of the length of a pathway given the numer of gene sets a gene is in.

condPerPathways: The conditional probability of the numer of gene sets a gene is in given the length of a gene set.

sizePathways

sizeGenes